requirements/base.txt
pytz==2020.1  # https://github.com/stub42/pytz
python-slugify==4.0.0  # https://github.com/un33k/python-slugify
Pillow==8.0  # https://github.com/python-pillow/Pillow
argon2-cffi==19.2.0  # https://github.com/hynek/argon2_cffi
whitenoise==5.0.1  # https://github.com/evansd/whitenoise
redis==3.5.0  # https://github.com/andymccurdy/redis-py
hiredis==1.0.1  # https://github.com/redis/hiredis-py

# Django
# ------------------------------------------------------------------------------
#django==3.0.5  # pyup: < 3.1  # https://www.djangoproject.com/
django-environ==0.4.5  # https://github.com/joke2k/django-environ
django-model-utils==4.0.0  # https://github.com/jazzband/django-model-utils
django-allauth==0.41.0  # https://github.com/pennersr/django-allauth
django-crispy-forms==1.13.0  # https://github.com/django-crispy-forms/django-crispy-forms
crispy-bootstrap5
django-redis==4.11.0  # https://github.com/jazzband/django-redis
django-user-language-middleware

# evolvemedicus
# ------------------------------------------------------------------------------
django-filter==2.4.0
django-modeltranslation==0.15
django-guardian==2.4.0
psycopg2-binary==2.8.5
django-rosetta==0.9.4
python-dateutil==2.8.1
django-extensions==2.2.9
django-embed-video==1.3.3
django-import-export==2.5.0
django-comments-xtd
geocoder==1.38.1
django-hitcount
django-resized

##data analysis
#scikit-learn==1.0.2
#pandas==1.4.2
#xlsxwriter